Abstracted GitHub out of the OAuth2 provider
