+++ build.gradle.kts
@@ -8,7 +8,7 @@ plugins {
 }
 
 group = "dev.castive"
-version = "0.5.0"
+version = "0.5.1"
 
 repositories {
 	mavenCentral()
